Add file-based dye tracer
This teaches MOM6 the "USE_FILE_DYES" option. If true, it enables a
new tracer package that is similar to the dye tracer with a box
source, except the source is defined from a file. The file is
specified with "FILE_DYE_TRACERS_FILE", and contains
"NUM_FILE_DYE_TRACERS" fields named "dyeN", where N ranges from 1 to
the number of tracers. These tracers experience vertical diffusion,
and are otherwise advected around as a concentration. The
concentration in the source region is constantly set to 1, providing a
continuous source.