Fix indexing i/j typo bug when HMIX_UV_SFC_PROP>0 (#28)
There was a typo with the calculation of mean velocities in the uppermost depth_ml fluid.
It was written J=js-1,ie instead of J=js-1,je and therefore crashed with an indexing error
when used. This section of the code only gets used when (CS%Hmix_UV>0.) i.e. when
runtime parameter HMIX_UV_SFC_PROP > 0. The runtime parameter default is zero.
I have not added a runtime parameter fixing the bug optionally because without the bugfix
it would crash if using this parameter. However, it should not change answers in simulations
that use the default HMIX_UV_SFC_PROP.