Add mom_cap_outputlog.F90 that enables output logging diagnostics at a given hourly output frequency
Author: Denise Worthen
This feature is required for UFS operational configurations and is used to determine when 
MOM6 output (diagnostics and restart) have been completed. The log files created by 
this feature can be queried by the Global Workflow to either trigger downstream jobs 
or to ensure that if a run fails and a restart is required, model output is available 
consistent with a given restart file.